feat: experimental alternate ways to compile typescript without babel

+++ packages/pob/lib/generators/common/babel/CommonBabelGenerator.js
@@ -1,4 +1,3 @@
-import fs from 'node:fs';
 import semver from 'semver';
 import Generator from 'yeoman-generator';
 import * as packageUtils from '../../../utils/package.js';
@@ -251,8 +250,10 @@ export default class CommonBabelGenerator extends Generator {
     delete pkg.pob.withReact;
     if (newBabelEnvs.length === 0) {
       delete pkg.pob.babelEnvs;
-      delete pkg.pob.entries;
-      delete pkg.pob.jsx;
+      if (!pkg.pob.typescript) {
+        delete pkg.pob.entries;
+        delete pkg.pob.jsx;
+      }
     } else {
       pkg.pob.babelEnvs = newBabelEnvs;
       pkg.pob.entries = pkg.pob.entries || ['index'];
@@ -270,10 +271,6 @@ export default class CommonBabelGenerator extends Generator {
     const pkg = this.fs.readJSON(this.destinationPath('package.json'));
     this.entries = pkg.pob.entries;
     this.babelEnvs = pkg.pob.babelEnvs || [];
-
-    if (this.babelEnvs.length > 0) {
-      fs.mkdirSync(this.destinationPath('src'), { recursive: true });
-    }
 
     if (this.entries) {
       this.entries.forEach((entry) => {
@@ -300,69 +297,15 @@ export default class CommonBabelGenerator extends Generator {
   }
 
   default() {
-    const useBabel = this.babelEnvs && this.babelEnvs.length > 0;
     const pkg = this.fs.readJSON(this.destinationPath('package.json'));
-    const hasTargetNode = this.babelEnvs.find((env) => env.target === 'node');
+    const useBabel = this.babelEnvs && this.babelEnvs.length > 0;
+    const useTypescript = useBabel || pkg.pob?.typescript;
+    const hasTargetNode = useBabel
+      ? this.babelEnvs.find((env) => env.target === 'node')
+      : useTypescript; // todo pkg.pob.typescriptTargets
     const hasTargetBrowser = this.babelEnvs.find(
       (env) => env.target === 'browser',
     );
-
-    /* scripts */
-
-    if (this.options.isApp) {
-      packageUtils.removeScripts(['watch']);
-      packageUtils.addOrRemoveScripts(pkg, useBabel, {
-        'clean:build': `pob-babel-clean-out ${this.options.buildDirectory}`,
-        build: 'yarn clean:build && rollup --config rollup.config.mjs',
-        start: 'yarn clean:build && rollup --config rollup.config.mjs --watch',
-        clean: 'yarn clean:build',
-      });
-    } else {
-      packageUtils.removeScripts(['start']);
-      packageUtils.addScripts(pkg, {
-        'clean:build': useBabel
-          ? `pob-babel-clean-out ${this.options.buildDirectory}`
-          : 'true',
-      });
-      packageUtils.addOrRemoveScripts(pkg, useBabel, {
-        build: 'yarn clean:build && rollup --config rollup.config.mjs',
-        watch: 'yarn clean:build && rollup --config rollup.config.mjs --watch',
-        clean: 'yarn clean:build',
-      });
-    }
-
-    const shouldBuildDefinitions = !this.options.isApp && useBabel;
-    packageUtils.addOrRemoveScripts(pkg, shouldBuildDefinitions, {
-      'build:definitions': 'tsc -p tsconfig.json',
-    });
-
-    if (shouldBuildDefinitions) {
-      pkg.scripts.build += ' && yarn run build:definitions';
-    } else if (!this.options.isApp && !useBabel) {
-      // check definitions, but also force lerna to execute build:definitions in right order
-      // example: nightingale-types depends on nightingale-levels
-      if (this.fs.exists(this.destinationPath('lib/index.d.ts'))) {
-        packageUtils.addScripts(pkg, {
-          'build:definitions':
-            'tsc --lib esnext --noEmit --skipLibCheck ./lib/index.d.ts',
-          build: 'yarn run build:definitions',
-        });
-      }
-
-      if (this.fs.exists(this.destinationPath('lib/index.ts'))) {
-        packageUtils.addScripts(pkg, {
-          'build:definitions':
-            'tsc --lib esnext --noEmit --skipLibCheck ./lib/index.ts',
-          build: 'yarn run build:definitions',
-        });
-      }
-    }
-
-    if (pkg.scripts) {
-      delete pkg.scripts.postbuild;
-      delete pkg.scripts['build:dev'];
-      delete pkg.scripts['watch:dev'];
-    }
 
     /* dependencies */
 
@@ -411,11 +354,14 @@ export default class CommonBabelGenerator extends Generator {
 
     if (hasTargetNode) {
       if (!pkg.engines) pkg.engines = {};
-      const minNodeVersion = Math.min(
-        ...this.babelEnvs
-          .filter((env) => env.target === 'node')
-          .map((env) => env.version),
-      );
+      const minNodeVersion = useBabel
+        ? Math.min(
+            ...this.babelEnvs
+              .filter((env) => env.target === 'node')
+              .map((env) => env.version),
+          )
+        : // eslint-disable-next-line unicorn/no-unreadable-iife
+          (() => (this.options.onlyLatestLTS ? '20' : '18'))();
       switch (String(minNodeVersion)) {
         case '10':
         case '12':
@@ -492,245 +438,6 @@ export default class CommonBabelGenerator extends Generator {
       delete pkg.browserslist;
     }
 
-    /* side effects */
-
-    if (this.options.isApp && !this.options.isAppLibrary) {
-      delete pkg.sideEffects;
-    } else if (!('sideEffects' in pkg)) {
-      pkg.sideEffects = true;
-      console.warn('Setting pkg.sideEffects to true, as it was not defined');
-    } else if (pkg.sideEffects) {
-      console.warn(
-        "pkg.sideEffects is true, are you sure you can't set it to false ?",
-      );
-    }
-
-    /* main / aliases / typing */
-
-    if (this.options.isApp) {
-      delete pkg.types;
-      delete pkg.typings;
-    } else if (pkg.typings) {
-      if (!pkg.types) pkg.types = pkg.typings;
-      delete pkg.typings;
-    }
-
-    // if (!pkg.main || pkg.main.startsWith('./lib/')) {
-    if (useBabel) {
-      // see pkg.exports instead.
-      delete pkg.main;
-      if (!this.options.isApp) {
-        pkg.types = `./${this.options.buildDirectory}/definitions/index.d.ts`;
-      }
-    } else {
-      if (!pkg.main) {
-        pkg.exports = './lib/index.js';
-      }
-      if (pkg.type === 'module' && this.fs.exists('./lib/index.cjs')) {
-        pkg.main = './lib/index.cjs';
-      } else {
-        pkg.main = './lib/index.js';
-      }
-      if (!this.options.isApp) {
-        if (this.fs.exists('./lib/index.ts')) {
-          pkg.types = './lib/index.ts';
-        } else if (this.fs.exists('./lib/index.d.ts') || pkg.types) {
-          pkg.types = './lib/index.d.ts';
-        }
-      }
-      if (!pkg.engines) pkg.engines = {};
-      if (
-        !pkg.engines.node ||
-        semver.lt(
-          semver.minVersion(pkg.engines.node),
-          // pkg.type === 'commonjs' ? '12.10.0' : '12.20.0',
-          '14.13.1',
-        )
-      ) {
-        pkg.engines.node = '>=16.0.0';
-      }
-    }
-
-    delete pkg['browser-dev'];
-    delete pkg['module-dev'];
-
-    const esAllBrowserEnv = this.babelEnvs.find(
-      (env) =>
-        env.target === 'browser' &&
-        env.version === undefined &&
-        (!env.formats || env.formats.includes('es')),
-    );
-
-    // Legacy "dev" builds
-    delete pkg['module:browser'];
-    delete pkg['module:browser-dev'];
-    delete pkg['module:modern-browsers'];
-    delete pkg['module:modern-browsers-dev'];
-    delete pkg['module:node'];
-    delete pkg['module:node-dev'];
-
-    /* webpack 4 */
-    if (esAllBrowserEnv) {
-      pkg.module = `./${this.options.buildDirectory}/index-browser.es.js`;
-      pkg.browser = `./${this.options.buildDirectory}/index-browser.es.js`;
-    } else {
-      delete pkg.module;
-      delete pkg.browser;
-    }
-
-    /* webpack 5 and node with ESM support */
-    if (useBabel) {
-      pkg.exports = {
-        './package.json': './package.json',
-      };
-
-      this.entries.forEach((entry) => {
-        const isBrowserOnly =
-          entry === 'browser' &&
-          (this.babelEnvs.every((env) => env.target === 'browser') ||
-            (this.entries.length === 2 && this.entries.includes('index')));
-        const entryDistName = isBrowserOnly ? 'index' : entry;
-        const exportName = entry === 'index' ? '.' : `./${entry}`;
-
-        const targets = {
-          types:
-            pkg.private || this.options.isAppLibrary
-              ? `./src/${entryDistName}.ts`
-              : `./${this.options.buildDirectory}/definitions/${entryDistName}.d.ts`,
-        };
-
-        const defaultNodeEnv = this.babelEnvs.find(
-          (env) => env.target === 'node',
-        );
-        const defaultNodeEnvVersion = defaultNodeEnv && defaultNodeEnv.version;
-
-        this.babelEnvs.forEach(({ target, version, formats }) => {
-          if (target === 'node' && entry === 'browser') return;
-
-          const exportTarget = {};
-
-          if (target === 'node') {
-            const cjsExt = pkg.type === 'module' ? 'cjs' : 'cjs.js';
-            if (!formats || formats.includes('es')) {
-              exportTarget.import = `./${this.options.buildDirectory}/${entryDistName}-${target}${version}.mjs`;
-
-              if (formats && formats.includes('cjs')) {
-                exportTarget.require = `./${this.options.buildDirectory}/${entryDistName}-${target}${version}.${cjsExt}`;
-              }
-            } else if (formats && formats.includes('cjs')) {
-              exportTarget.default = `./${this.options.buildDirectory}/${entryDistName}-${target}${version}.${cjsExt}`;
-            }
-            // eslint: https://github.com/benmosher/eslint-plugin-import/issues/2132
-            // jest: https://github.com/facebook/jest/issues/9771
-            if (!pkg.main && exportName === '.') {
-              pkg.main =
-                pkg.type === 'module'
-                  ? exportTarget.import
-                  : exportTarget.default ||
-                    exportTarget.require ||
-                    exportTarget.import;
-            }
-          } else if (target === 'browser') {
-            if (!formats || formats.includes('es')) {
-              exportTarget.import = `./${
-                this.options.buildDirectory
-              }/${entryDistName}-${target}${version || ''}.es.js`;
-            }
-
-            if (formats && formats.includes('cjs')) {
-              exportTarget.require = `./${
-                this.options.buildDirectory
-              }/index-${target}${version || ''}.cjs.js`;
-            }
-          }
-
-          if (
-            !version ||
-            (target === 'node' && version === defaultNodeEnvVersion)
-          ) {
-            targets[target] = {
-              ...targets[target],
-              ...exportTarget,
-            };
-          } else {
-            targets[target] = {
-              [`${target}:${version}`]: exportTarget,
-              ...targets[target],
-            };
-          }
-        });
-
-        pkg.exports[exportName] = targets;
-      });
-
-      if (pkg.pob.extraEntries) {
-        pkg.pob.extraEntries.forEach((extraEntryConfig) => {
-          if (typeof extraEntryConfig === 'string') {
-            extraEntryConfig = {
-              name: extraEntryConfig,
-            };
-          }
-
-          const calcExport = () => {
-            if (pkg.type === 'module') {
-              return extraEntryConfig.name.endsWith('.cjs') ||
-                extraEntryConfig.name.endsWith('.d.ts')
-                ? `./${extraEntryConfig.name}`
-                : `./${extraEntryConfig.name}.js`;
-            }
-
-            return {
-              import: `./${extraEntryConfig.name}.mjs`,
-              require: `./${extraEntryConfig.name}.js`,
-            };
-          };
-
-          let exportValue = calcExport();
-
-          if (extraEntryConfig.types) {
-            if (typeof exportValue === 'string') {
-              exportValue = {
-                types: `./${extraEntryConfig.types}`,
-                default: exportValue,
-              };
-            } else {
-              exportValue = {
-                types: `./${extraEntryConfig.types}`,
-                ...exportValue,
-              };
-            }
-          }
-
-          pkg.exports[`./${extraEntryConfig.name}`] = exportValue;
-        });
-      }
-    } else if (!pkg.exports) {
-      console.error('Please setup your package.exports manually.');
-    } else {
-      if (typeof pkg.exports === 'string') {
-        pkg.exports = {
-          '.': pkg.exports,
-        };
-      }
-      if (!pkg.exports['./package.json']) {
-        pkg.exports['./package.json'] = './package.json';
-      }
-
-      if (pkg.types && !pkg.exports['.'].types) {
-        if (typeof pkg.exports['.'] === 'string') {
-          pkg.exports['.'] = {
-            default: pkg.exports['.'],
-          };
-        }
-        pkg.exports['.'] = { types: pkg.types, ...pkg.exports['.'] };
-      }
-    }
-
-    Object.keys(pkg).forEach((key) => {
-      if (!key.startsWith('module:') && !key.startsWith('webpack:')) return;
-      delete pkg[key];
-    });
-
     this.fs.writeJSON(this.destinationPath('package.json'), pkg);
   }
 
@@ -778,7 +485,7 @@ export default class CommonBabelGenerator extends Generator {
           },
         );
       }
-    } else {
+    } else if (!pkg.pob.typescript && pkg.pob.rollup !== false) {
       this.fs.delete('rollup.config.mjs');
     }
